// src/helpers/calculateTiming.d.ts
export declare const displayPeriod: (startDate: string, duration: string) => string;
export declare const displayDuration: (startDate: string, duration: string | number) => string;
export declare const getPostingSchedule: (numPosts: number) => string;
export declare const getPostingScheduleArray: (numPosts: number) => string[];